\name{format_legend.aloha}
\alias{format_legend}
\alias{format_legend.aloha}
\alias{format_legend.amdahl}
\alias{format_legend.gustafson}
\alias{format_legend.sfit}
\title{Format a legend for a specific object}
\usage{
  \method{format_legend}{aloha} (object)

  format_legend(object)

  \method{format_legend}{gustafson} (object)

  \method{format_legend}{amdahl} (object)

  \method{format_legend}{sfit} (object)
}
\arguments{
  \item{object}{a sfit object}
}
\description{
  S3 method format a legend in plot.sfit based on the
  sfit_type
}

