\docType{data}
\name{ngnix}
\alias{ngnix}
\title{Example 4 of scaling data for ngnix server}
\format{A data frame with 9 rows and 2 variables}
\description{
  An example of scaling data for ngnix web server
}
\details{
  \itemize{ \item N. 9 user levels \item X_N. corresponding
  measured throughput }
}
\keyword{datasets}

