\name{Nmax}
\alias{Nmax}
\alias{Nmax.sfit}
\title{Get Max # of users for fit attribute from an sfit object}
\usage{
  Nmax(object)

  \method{Nmax}{sfit} (object)

  \method{Nmax}{sfit} (object)
}
\arguments{
  \item{object}{a sfit object}
}
\description{
  S3 method to access sfit object's sfit_type attribute
}

