\docType{data}
\name{postgreSQL}
\alias{postgreSQL}
\title{Example 2 of scaling data}
\format{A data frame with 21 rows and 2 variables}
\description{
  A example of data for postgreSQL analysis
}
\details{
  \itemize{ \item N. 21 client levels \item X_N.
  corresponding measured throughput }
}
\keyword{datasets}

