\name{sfit_type}
\alias{sfit_type}
\alias{sfit_type.sfit}
\title{Get sfit sfit_type attribute from an sfit object}
\usage{
  sfit_type(object)

  \method{sfit_type}{sfit} (object)
}
\arguments{
  \item{object}{a sfit object}
}
\description{
  S3 method to access sfit object's sfit_type attribute
}

