\docType{data}
\name{SSUSLcalc}
\alias{SSUSLcalc}
\title{Example 3 of scaling data}
\format{A data frame with 7 rows and 2 variables}
\description{
  A unstructure dataset containing user levels and measured
  throughputs
}
\details{
  \itemize{ \item V1. 7 user levels \item V2. corresponding
  measured throughput }
}
\keyword{datasets}

