\docType{data}
\name{USLcalc}
\alias{USLcalc}
\title{Example 1 of scaling data}
\format{A data frame with 11 rows and 2 variables}
\description{
  A dataset containing user levels and measured throughputs
}
\details{
  \itemize{ \item N. 11 user levels \item X_N.
  corresponding measured throughput }
}
\keyword{datasets}

