\name{zones_to_frame}
\alias{zones_to_frame}
\alias{zones_to_frame.zones}
\title{Convert zones object to data.frame}
\usage{
  zones_to_frame(object, ...)

  \method{zones_to_frame}{zones} (object,
    predict_vector = NULL)
}
\arguments{
  \item{object}{zones objects}

  \item{object}{a zones}

  \item{object}{zones object}
}
\description{
  Convert zones object to data.frame
}

