\name{zones_to_frame}
\alias{zones_to_frame}
\alias{zones_to_frame.zones}
\title{Convert zones object to data.frame}
\usage{
  zones_to_frame(object, ...)

  \method{zones_to_frame}{} (object, predict_vector = NULL)
}
\arguments{
  \item{object}{a zones}
}
\description{
  Convert zones object to data.frame
}

